+++ source/dlp/commands/utility.d
@@ -56,14 +56,32 @@ bool hasErrors(const ref Global global)
     return global.errors > 0;
 }
 
+void addStringImport(const(char)[] path)
+{
+    import std.string : toStringz;
+
+    import dmd.globals : global;
+    import dmd.arraytypes : Strings;
+
+    if (global.filePath is null)
+        global.filePath = new Strings();
+
+    if (global.params.fileImppath is null)
+        global.params.fileImppath = new Strings();
+
+    global.filePath.push(path.toStringz);
+    global.params.fileImppath.push(path.toStringz);
+}
+
 Module runFullFrontend(Ast = ASTCodegen)(
     const string filename,
     const string content,
-    const string[] importPaths)
+    const string[] importPaths,
+    const string[] stringImportPaths)
 {
 
     return runParser!Ast(filename, content, importPaths)
-        .runSemanticAnalyzer();
+        .runSemanticAnalyzer(stringImportPaths);
 }
 
 Module runParser(Ast = ASTCodegen)(
@@ -93,9 +111,13 @@ Module runParser(Ast = ASTCodegen)(
     return t.module_;
 }
 
-Module runSemanticAnalyzer(Module module_)
+Module runSemanticAnalyzer(Module module_, const string[] stringImportPaths)
 {
+    import std.algorithm : each;
+
     import dmd.frontend : fullSemantic;
+
+    stringImportPaths.each!addStringImport;
 
     fullSemantic(module_);
     handleDiagnosticErrors();